Refeitas lógicas do banco de dados, alterando para banco relacional e não mais orientado a objetos

+++ ChatFodaPraCaralho/src/chat/foda/pra/caralho/rmi/ClienteRemoto.java
@@ -3,7 +3,7 @@ package chat.foda.pra.caralho.rmi;
 import java.rmi.Remote;
 import java.rmi.RemoteException;
 
-import chat.foda.pra.caralho.modelo.Chat;
+import chat.foda.pra.caralho.models.Chat;
 
 /**
  * Declaração de todos os métodos executados no Cliente visíveis para o Servidor
@@ -11,14 +11,14 @@ import chat.foda.pra.caralho.modelo.Chat;
  * @author luiznazari
  */
 public interface ClienteRemoto extends Remote {
-
-	public void enviarMensagem(Integer chatCodigo, String mensagem) throws RemoteException;
+	
+	public void enviarMensagem(Long chatCodigo, String mensagem) throws RemoteException;
 	
 	public void enviarParaTodos(String mensagem) throws RemoteException;
 	
 	public void abrirChat(Chat chat, String nomeAmigo) throws RemoteException;
 	
-	public void desativarChat(Integer chatCodigo) throws RemoteException;
+	public void desativarChat(Long chatCodigo) throws RemoteException;
 	
 	public void desativarTodosChats() throws RemoteException;
