+++ ChatFodaPraCaralho/src/chat/foda/pra/caralho/rmi/ServidorRemoto.java
@@ -4,9 +4,9 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 import java.util.ArrayList;
 
-import chat.foda.pra.caralho.modelo.Chat;
-import chat.foda.pra.caralho.modelo.Usuario;
-import chat.foda.pra.caralho.modelo.UsuarioLogado;
+import chat.foda.pra.caralho.models.Chat;
+import chat.foda.pra.caralho.models.Usuario;
+import chat.foda.pra.caralho.models.UsuarioLogado;
 
 /**
  * Declaração de todos os métodos executados no servidor visíveis para o Cliente
@@ -21,9 +21,9 @@ public interface ServidorRemoto extends Remote {
 	
 	public void removerUsuario(String nome) throws RemoteException;
 	
-	public void enviarMensagemParaServidor(Integer chatCodigo, String mensagem) throws RemoteException;
+	public void enviarMensagemParaServidor(Long chatCodigo, String mensagem) throws RemoteException;
 	
-	public void logout(ArrayList<Integer> codigos, ClienteRemoto cliente, String nome) throws RemoteException;
+	public void logout(ArrayList<Long> codigos, ClienteRemoto cliente, String nome) throws RemoteException;
 	
 	public boolean adicionaAmigo(Usuario usuario, String nomeAmigo) throws RemoteException;
